requirements.txt
absl==0.0
absl_py==1.0.0
clu==0.0.6
flax==0.4.1
jax==0.3.4
jax_dataclasses==1.2.1
jaxfg.egg==info
matplotlib==3.5.1
ml_collections==0.1.1
numpy==1.21.5
optax==0.1.1
overrides==6.1.0
seaborn==0.11.2
tensorflow==2.8.0
